Projects/requirements.txt
langchain==0.0.208
deeplake==3.6.5
openai==0.27.8
tiktoken==0.4.0
elevenlabs==0.2.18
streamlit==1.23.1
beautifulsoup4==4.11.2
audio-recorder-streamlit==0.0.8
streamlit-chat==0.0.2.2
langchain-google-genai
python-dotenv